requirements.txt
python-dotenv
psycopg2-binary
SPARQLWrapper
requests
fastapi
uvicorn